requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Werkzeug==2.3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.6
itsdangerous==2.1.2
blinker==1.6.2
